requirements.txt
discord.py-self
python-dotenv
groq
openai
aiohttp
colorama
